src/tests/e2e/ui/requirements.txt
# Copyright (C) 2024-2026 Intel Corporation
# SPDX-License-Identifier: Apache-2.0

aiohttp==3.12.15
pytest==7.4.4 # MIT
pytest-asyncio==0.23.8
playwright==1.48.0
pytest-playwright==0.5.2
allure-pytest==2.13.5
python-dotenv==1.0.0
sniffio==1.3.1